users/urls.py
from django.urls import path
from rest_framework.routers import DefaultRouter
from .viewsets import UserViewSet
from rest_framework_simplejwt.views import (
    TokenObtainPairView,
    TokenRefreshView,
)

router = DefaultRouter()
router.register("users", UserViewSet, basename='user')

users_urlpatterns = router.urls

auth_urlpatterns = [
    path("login/", TokenObtainPairView.as_view(), name="token_obtain_pair"),
    path("token/refresh/", TokenRefreshView.as_view(), name="token_refresh"),
]
